Add argument handling to cli
This change makes it possible to add arguments to the command line
interface. At this stage, only recognized option is -f [file name] for
specifying the state file name. For example, use
    python pet_cli.py -f state.json
to store the data to the file state.json.

Arguments are parsed in reverse order using a stack machine style code.
Any argument not recognized as an option will be considered an operand
and pushed on the stack. Option flags work like operators. They pop the
expected number of arguments and perform whatever operation they need.
Adding more options is easy, just add an elif for that option.